openrouter_ai/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
streamlit
tqdm
requests
tiktoken
pytest
scikit-learn
